+++ NEWS.md
@@ -1,3 +1,9 @@
+# MazamaSatelliteUtils 0.2.12
+
+ * `goesaodc_createNativeGrid()` now properly reads and scales the raw data 
+ which are stored as `unsigned short int` values.
+ * New `goesaodc_areaPlot()` function creates nice plots of AOD.
+
 # MazamaSatelliteUtils 0.2.11
 
 * `goesaodc_downloadAOD()` and `goesaodc_downloadDaytimeAOD()` now return a